requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
structlog==24.1.0
requests==2.31.0
python-multipart==0.0.6
groq>=0.4.0
google-generativeai>=0.8.0
PyPDF2
# RAG (Heavy - might exceed Vercel limits)
chromadb
sentence-transformers
